Half a cycle working

+++ src/motor.h
@@ -6,14 +6,14 @@ class Motor {
         Motor(int forwardPin, int reversePin);
         void run(int speed);
         void stop();
-        void setAccel(int creepPower, int maxPower, int acceleration, double anticipatedTime);
+        void setAccel(int creepPower, int maxPower, double acceleration, double anticipatedTime);
         void runAccel();
     private:
         int _forwardPin;
         int _reversePin;
         int _creepPower;
         int _maxPower;
-        int _acceleration;
+        double _acceleration;
         double _anticipatedTime;
         unsigned long motorStartTime;
 };
